requirements.txt
streamlit
tensorflow
pillow
opencv-python
numpy
mysql-connector-python
